+++ IDP/PersonalFinanceTracker/src/main/java/com/arijit/idp/service/IncomeServiceImpl.java
@@ -7,6 +7,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.arijit.idp.entity.Income;
+import com.arijit.idp.exception.IncomeAlreadyPresentException;
+import com.arijit.idp.exception.IncomeNotFoundException;
+import com.arijit.idp.exception.InvalidDataFormatException;
+import com.arijit.idp.exception.NotAStringException;
+import com.arijit.idp.exception.NullValueEnteredException;
 import com.arijit.idp.repository.IncomeRepository;
 
 @Service
@@ -16,28 +21,85 @@ public class IncomeServiceImpl implements IncomeService {
 	private IncomeRepository incomeRepository;
 
 	// Create
-	public Income insertIncome(Income income) {
+	@Override
+	public Income insertIncome(Income income) throws IncomeAlreadyPresentException {
+
+		if (incomeRepository.existsById(income.getIncomeId())) {
+			throw new IncomeAlreadyPresentException();
+		}
 		return incomeRepository.save(income);
 	}
 
 //	// Retrieve
-	public List<Income> findByUserId(String userId) {
+	@Override
+	public List<Income> findByUserId(String userId)
+			throws IncomeNotFoundException, NullValueEnteredException, NotAStringException {
+
+		if (userId == null || userId.isEmpty()) {
+			throw new NullValueEnteredException("User Id cannot be empty");
+		}
+		if (!(userId instanceof String)) {
+			throw new NotAStringException("The entered password is not a String value");
+		}
 		List<Income> incomeByUserId = incomeRepository.findByUserId(userId);
+		if (incomeByUserId.isEmpty()) {
+			throw new IncomeNotFoundException();
+		}
 		return incomeByUserId;
 	}
+	
+	@Override
+	public List<Income> findByUserIdAndMonth(String userId, int month)
+			throws NullValueEnteredException, NotAStringException, InvalidDataFormatException, IncomeNotFoundException {
 
-	public List<Income> findByUserIdAndMonth(String userId, int month) {
+		if (userId == null || userId.isEmpty()) {
+			throw new NullValueEnteredException("User Id cannot be empty");
+		}
+		if (!(userId instanceof String)) {
+			throw new NotAStringException("The entered password is not a String value");
+		}
+		if ((!(Integer.valueOf(month) instanceof Integer)) || month < 1 || month > 12) {
+			throw new InvalidDataFormatException("Month should be between 1 to 12");
+		}
 		List<Income> incomeByUserIdAndMonth = incomeRepository.findByUserIdAndMonth(userId, month);
+		if (incomeByUserIdAndMonth == null) {
+			throw new IncomeNotFoundException();
+		}
 		return incomeByUserIdAndMonth;
 	}
 
-	public List<Income> findByUserIdAndYear(String userId, int year) {
+	@Override
+	public List<Income> findByUserIdAndYear(String userId, int year)
+			throws NullValueEnteredException, NotAStringException, InvalidDataFormatException, IncomeNotFoundException {
+
+		if (userId == null || userId.isEmpty()) {
+			throw new NullValueEnteredException("User Id cannot be empty");
+		}
+		if (!(userId instanceof String)) {
+			throw new NotAStringException("The entered password is not a String value");
+		}
+		if ((!(Integer.valueOf(year) instanceof Integer)) || year < 1000 || year > 9999) {
+			throw new InvalidDataFormatException("Invalid year entered");
+		}
+
 		List<Income> incomeByUserIdAndYear = incomeRepository.findByUserIdAndYear(userId, year);
+		if (incomeByUserIdAndYear == null) {
+			throw new IncomeNotFoundException();
+		}
 		return incomeByUserIdAndYear;
 	}
 
 	// Update
-	public Income updateByIncomeId(int incomeId, Income updatedIncome) {
+	@Override
+	public Income updateByIncomeId(int incomeId, Income updatedIncome)
+			throws NullValueEnteredException, InvalidDataFormatException, IncomeNotFoundException {
+
+		if (incomeId == 0) {
+			throw new NullValueEnteredException("Income Id cannot be empty");
+		}
+		if (!(Integer.valueOf(incomeId) instanceof Integer)) {
+			throw new InvalidDataFormatException();
+		}
 		Optional<Income> existingIncome = incomeRepository.findById(incomeId);
 		Income newIncome = null;
 		if (existingIncome.isPresent()) {
@@ -45,22 +107,29 @@ public class IncomeServiceImpl implements IncomeService {
 			newIncome.setIncomeDate(updatedIncome.getIncomeDate());
 			newIncome.setIncomeAmount(updatedIncome.getIncomeAmount());
 			incomeRepository.save(newIncome);
+		} else {
+			throw new IncomeNotFoundException("No income found with given income id");
 		}
 		return newIncome;
 	}
 
 	// Delete
-	public String deleteByIncomeId(int incomeId) {
-		String status;
+	@Override
+	public void deleteByIncomeId(int incomeId)
+			throws NullValueEnteredException, InvalidDataFormatException, IncomeNotFoundException {
+
+		if (incomeId == 0) {
+			throw new NullValueEnteredException("Income Id cannot be empty");
+		}
+		if (!(Integer.valueOf(incomeId) instanceof Integer)) {
+			throw new InvalidDataFormatException();
+		}
 		Optional<Income> existingIncome = incomeRepository.findById(incomeId);
-		if(existingIncome.isPresent()) {
+		if (existingIncome.isPresent()) {
 			incomeRepository.deleteById(incomeId);
-			status = "Income removed for Income Id: " + incomeId;
+		} else {
+			throw new IncomeNotFoundException("No income found with given income id");
 		}
-		else {
-			status = "Income not found"; //Add exception handling
-		}
-		return status;
 	}
 
 }
